Aquaterm 1.0.1
A Mac OS X plotting front end (grahics terminal) written using
Objective-C and the Cocoa API.

It allows other applications to display vector graphics and images using
a simple API. This is useful for displaying graphics generated by old
code natively in Aqua. AquaTerm will let you print and save the output
as PDF or EPS files.

You connect to AquaTerm through an "adapter", a small code snippet that
will have to be linked with your old code. If you are lucky, an adapter
is already available for download, otherwise you'll have to write your
own adapter using one of the examples provided.

Our aim is to provide as many adapters as possible for command line UNIX
programs and scripting languages and we would be happy to host any
efforts to write new adapters.

At present, there are adapters for gnuplot, PGPLOT and PLplot. AquaTerm
can be used with C, Objective-C, FORTRAN, Perl, and Python.