New Formula: mscgen
A small program that parses Message Sequence Chart descriptions and
produces PNG, SVG, EPS or server side image maps (ismaps) as the output.

Mscgen aims to provide a simple text language that is clear to create, edit
and understand, which can also be transformed into common image formats for
display or printing.